requirements.txt
Flask
Flask-Cors
Pillow
pytesseract
easyocr